moved table to separate component
